refactor: dataZoom
